fix(pipeline): emit ChapterStart so TUI leaves init phase
After the split of app/workflows/project.py into pipeline.py + chapter.py
(commit 2d64d09), the ChapterStart event lost its producer — the scanner
stage no longer emitted it. TUI state stayed at phase=init forever, so
all three phase lines rendered 'waiting' while the pipeline was actually
running.

- pipeline.py: emit ChapterStart at the start of each scanner job, before
  engine.preprocess(). Consumers (RichHook, TUI) now transition init ->
  preprocess and display the real chapter number + page count.
- events.py: widen ChapterStart.chapter and ChapterSkipped.chapter from
  int to float to match ChapterDone.chapter and the actual float-valued
  chapter numbers (e.g. 1.5) flowing through the pipeline.

+++ typoon/app/events.py
@@ -120,7 +120,7 @@ class PageRendered(Event):
 @dataclass
 class ChapterStart(Event):
     project_id: int = 0
-    chapter: int = 0
+    chapter: float = 0
     pages: int = 0
 
 
@@ -134,7 +134,7 @@ class ChapterDone(Event):
 
 @dataclass
 class ChapterSkipped(Event):
-    chapter: int = 0
+    chapter: float = 0
     reason: str = ""
